changed build system to gradle
