+++ fusion-client/build.gradle
@@ -1,0 +1,33 @@
+plugins {
+    id 'java-library'
+    id 'maven-publish'
+    id 'signing'
+}
+
+group 'com.datameshgroup.fusion'
+version '1.0.0'
+
+repositories {
+    mavenCentral()
+}
+
+java {
+    withJavadocJar()
+    withSourcesJar()
+}
+
+compileJava{
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+dependencies {
+    testImplementation 'junit:junit:4.+'
+    implementation 'org.glassfish.tyrus.bundles:tyrus-standalone-client:1.9'
+    implementation 'com.squareup.moshi:moshi:1.12.0'
+    implementation 'com.datameshgroup.fusion:fusion-sdk:1.0.2'
+}
+
+test {
+    useJUnit()
+}
