--- fusion-client/pom.xml
@@ -1,36 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>com.dmg</groupId>
-	<artifactId>fusion-client</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-
-	<name>Fusion Client</name>
-	<description>Client which enables connection to the DataMesh Unify payments platform</description>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.squareup.moshi</groupId>
-			<artifactId>moshi</artifactId>
-			<version>1.12.0</version>
-		</dependency>
-		<dependency>
-		    <groupId>org.glassfish.tyrus.bundles</groupId>
-		    <artifactId>tyrus-standalone-client</artifactId>
-		    <version>1.9</version>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.13.1</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-	<properties>
-		<maven.compiler.source>1.7</maven.compiler.source>
-		<maven.compiler.target>1.7</maven.compiler.target>
-		<failOnMissingWebXml>false</failOnMissingWebXml>
-	</properties>
-</project>
