+++ fusion-client/src/main/java/au/com/dmg/fusion/util/MessageHeaderUtil.java
@@ -1,0 +1,17 @@
+package au.com.dmg.fusion.util;
+
+import java.security.SecureRandom;
+
+public class MessageHeaderUtil {
+
+	private static final String CHARS = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+	private static SecureRandom random = new SecureRandom();
+
+	public static String generateServiceID(int len) {
+		StringBuilder sb = new StringBuilder(len);
+		for (int i = 0; i < len; i++)
+			sb.append(CHARS.charAt(random.nextInt(CHARS.length())));
+		return sb.toString();
+	}
+
+}
